Split AST callable declaration collection
