+++ docs/PHASE_PLAN.md
@@ -101,6 +101,9 @@ checked-in docs, tests, and commits only.
 - Monomorphization now keeps generic struct/enum specialization and nested
   specialized type-ref emission in a focused module, separate from callable
   specialization and substitution helpers.
+- AST declaration collection now keeps callable task dispatch and generic
+  function/method template collection in a focused module, separate from import,
+  impl-block, type, behavior, and precollection work.
 - Resolver-backed callable type-reference validation now shares the collected
   signature/body validation helper after function, top-level method, and
   `Type.impl` method dispatch restore the resolver-owned callable key, with
